pdf-rag-chatbot/requirements.txt
streamlit==1.37.0
pypdf==4.3.1
sentence-transformers==3.0.1
chromadb==0.5.5
rank-bm25==0.2.2
torch>=2.2.0
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
transformers==4.44.2
python-dotenv>=1.0.1
together==1.3.10
